Move ExceptionUtils stuff to static methods
